+++ dashboard/utils.py
@@ -1,6 +1,10 @@
+from collections import Counter
+import string
 import pandas as pd
 import plotly.express as px
 import plotly
+import nltk
+from nltk.corpus import stopwords
 
 from constants import std_time, review_count, avg_time, median_time, pre_emotions_column, during_emotions_column, post_emotions_column, project_review_col, rubric_heading, satisfaction_mapping, satisfaction_colors
 
@@ -185,3 +189,33 @@ def create_sei_fig(sei_data: pd.DataFrame) -> plotly.graph_objs.Figure:
     )
     sei_fig.for_each_annotation(lambda a: a.update(text=a.text.split("=")[-1]))
     return sei_fig
+
+def create_sei_comment_fig(sei_comments: pd.DataFrame) -> plotly.graph_objs.Figure:
+    # Installs needed corpus data
+    nltk.download('punkt')
+    nltk.download('stopwords')
+    
+    # Tokenizes the comments and computes their counts
+    results = Counter()
+    sei_comments["Comment"].str.lower().apply(nltk.word_tokenize).apply(results.update)
+    word_counts = pd.DataFrame.from_dict(results, orient="index").reset_index()
+    word_counts = word_counts.rename(columns={"index": "Word", 0:"Count"}) 
+    
+    # Removes stop words and punctuation from the totals
+    stop = stopwords.words("english")
+    word_counts = word_counts[~word_counts["Word"].isin(stop)]
+    word_counts = word_counts[~word_counts["Word"].isin(list(string.punctuation))]
+    word_counts = word_counts[~word_counts["Word"].str.contains("'")]
+    
+    # Sorts and pulls the top 25 words
+    word_counts = word_counts.sort_values(by="Count", ascending=False)
+    word_counts = word_counts.head(25)
+    word_counts = word_counts.sort_values(by="Count")
+    sei_comment_fig = px.bar(
+        word_counts,
+        x="Count",
+        y="Word",
+        height=1000,
+        title="Top 25 Most Common Words in SEI Comments"
+    )
+    return sei_comment_fig
